requirements.txt
asgiref==3.3.1
attrs==21.4.0
black==22.1.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.4
debugpy==1.5.1
Django==3.1.7
django-createsuperuserwithpassword==2.0.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
Faker==13.3.2
idna==2.10
iniconfig==1.1.1
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathspec==0.9.0
pika==1.2.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.1
requests==2.25.1
responses==0.20.0
six==1.16.0
sqlparse==0.2.4
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.4
